Update .gitignore and enhance README for PaperTrader project
- Added archive file types (*.tar, *.tar.gz, *.zip) to .gitignore to prevent unnecessary files from being tracked.
- Expanded README to provide a comprehensive overview of the PaperTrader platform, detailing features, tech stack, architecture, and setup instructions.
- Removed outdated sections and improved clarity for better user understanding.
- Updated project structure and deployment instructions to reflect recent changes in the application architecture.
