--- frontend/src/components/Markets.js
@@ -1,3 +1,0 @@
-Here I want to show a list of the 10 most popular stocks 5 
-most popular index funds and 5 most popular crypto  where the user can and also see quick stats like is the stock up today by how much 
-click on the stock and see the details of the stock.
