+++ frontend/src/components/common/Stock.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+
+/**
+ * Stock detail component
+ * 
+ * TODO: Implement stock detail page with:
+ * - In-depth stock information
+ * - Performance graph over time
+ * - Historical data visualization
+ * - Key metrics and statistics
+ */
+const Stock: React.FC = () => {
+  return (
+    <div style={{ marginTop: '60px' }}>
+      <div className="card">
+        <h2>Stock Details</h2>
+        <p style={{ color: '#666' }}>
+          Stock detail page coming soon. This will show:
+        </p>
+        <ul>
+          <li>In-depth stock information</li>
+          <li>Performance graph over time</li>
+          <li>Historical data visualization</li>
+          <li>Key metrics and statistics</li>
+        </ul>
+      </div>
+    </div>
+  );
+};
+
+export default Stock;
+
