--- frontend/src/setupProxy.js
@@ -1,11 +1,0 @@
-const { createProxyMiddleware } = require('http-proxy-middleware');
-
-module.exports = function(app) {
-  app.use(
-    '/api',
-    createProxyMiddleware({
-      target: process.env.BACKEND_URL || 'http://backend:8080',
-      changeOrigin: true,
-    })
-  );
-};
